kryo package

+++ sansa-query/sansa-query-spark/src/test/scala/net/sansa_stack/query/spark/compliance/SPARQL11TestSuiteRunnerSpark.scala
@@ -1,17 +1,19 @@
 package net.sansa_stack.query.spark.compliance
 
-import scala.collection.JavaConverters._
-import com.holdenkarau.spark.testing.SharedSparkContext
+import com.holdenkarau.spark.testing.DataFrameSuiteBase
+import net.sansa_stack.query.spark.api.domain.{QueryEngineFactory, QueryExecutionFactorySpark}
+import net.sansa_stack.query.tests.W3CConformanceSPARQLQueryEvaluationTestSuiteRunner
+import net.sansa_stack.rdf.spark.utils.SparkSessionUtils
+import org.aksw.commons.sql.codec.util.SqlCodecUtils
 import org.apache.jena.graph.Triple
 import org.apache.jena.query.Query
 import org.apache.jena.rdf.model.Model
 import org.apache.jena.sparql.resultset.SPARQLResult
+import org.apache.spark.SparkConf
 import org.apache.spark.rdd.RDD
-import org.apache.spark.sql.SparkSession
-import org.scalatest.{ConfigMap, Suite}
-import net.sansa_stack.query.spark.api.domain.{QueryEngineFactory, QueryExecutionFactorySpark}
-import net.sansa_stack.query.tests.W3CConformanceSPARQLQueryEvaluationTestSuiteRunner
-import org.aksw.commons.sql.codec.util.SqlCodecUtils
+import org.scalatest.Suite
+
+import scala.collection.JavaConverters._
 
 /**
  * SPARQL 1.1 test suite runner on Apache Spark.
@@ -21,51 +23,26 @@ import org.aksw.commons.sql.codec.util.SqlCodecUtils
  */
 abstract class SPARQL11TestSuiteRunnerSpark
   extends W3CConformanceSPARQLQueryEvaluationTestSuiteRunner
-//
     with org.scalatest.BeforeAndAfterAllConfigMap
-    with SharedSparkContext { self: Suite =>
+    with DataFrameSuiteBase { self: Suite =>
 
-  System.setProperty("spark.serializer", "org.apache.spark.serializer.KryoSerializer")
-  System.setProperty("spark.kryo.registrator", String.join(
-    ", ",
-    "net.sansa_stack.rdf.spark.io.JenaKryoRegistrator",
-    "net.sansa_stack.query.spark.ontop.OntopKryoRegistrator"))
-  System.setProperty("spark.sql.crossJoin.enabled", "true")
+  override val invokeBeforeAllAndAfterAllEvenIfNoTestsAreExpected: Boolean = false
 
-  @transient private var _spark: SparkSession = _
-
-  lazy val spark = SparkSession.builder.config(
-    conf
+  override def conf: SparkConf = {
+    super.conf
       .set("spark.sql.crossJoin.enabled", "true")
       .set("spark.serializer", "org.apache.spark.serializer.KryoSerializer")
       .set("spark.kryo.registrator", String.join(
         ", ",
         "net.sansa_stack.rdf.spark.io.JenaKryoRegistrator",
         "net.sansa_stack.query.spark.ontop.OntopKryoRegistrator"))
-  ).getOrCreate()
-
-
-  override val invokeBeforeAllAndAfterAllEvenIfNoTestsAreExpected: Boolean = false
-
-  override def beforeAll(configMap: ConfigMap): Unit = {
-    conf.set("spark.sql.crossJoin.enabled", "true")
-      .set("spark.serializer", "org.apache.spark.serializer.KryoSerializer")
-      .set("spark.kryo.registrator", String.join(
-        ", ",
-        "net.sansa_stack.rdf.spark.io.JenaKryoRegistrator",
-        "net.sansa_stack.query.spark.ontop.OntopKryoRegistrator"))
-    _spark = SparkSession.builder.config(conf).master("local[1]").getOrCreate()
-
-    val toIgnore = configMap.get("ignore")
-    if (toIgnore.nonEmpty) {
-      toIgnore.get.asInstanceOf[String].split(",").map(_.trim).foreach(name => IGNORED_NAMES.add(name))
-    }
   }
 
-  override def afterAll(configMap: ConfigMap): Unit = {
+  override def beforeAll(): Unit = {
+    super.beforeAll()
+  }
+  override def afterAll(): Unit = {
     super.afterAll()
-    spark.stop()
-    _spark = null
   }
 
   def getEngineFactory: QueryEngineFactory
@@ -75,31 +52,35 @@ abstract class SPARQL11TestSuiteRunnerSpark
   val db = "TEST"
 
   var qef: QueryExecutionFactorySpark = _
+  var previousDataURL: String = _
+  var data: Model = _
+  var dataRDD: RDD[Triple] = _
 
   override def loadData(datasetURL: String): Model = {
-    val data = super.loadData(datasetURL)
+    if (datasetURL != previousDataURL) {
+      previousDataURL = datasetURL
 
-    // we drop the Spark database to remove all tables from previous loaded data
-    spark.sql(s"DROP DATABASE IF EXISTS $db")
-    val sqlEscaper = SqlCodecUtils.createSqlCodecForApacheSpark()
-    spark.catalog.listDatabases().collect().foreach(db => {
-      spark.catalog.listTables(db.name).collect().foreach(t => {
-        val b = spark.catalog.dropTempView(s"${sqlEscaper.forSchemaName().encode(t.name)}")
-      })
-    })
+      data = super.loadData(datasetURL)
 
-    // we cannot handle empty data
-    if (!data.isEmpty) {
-      // distribute on Spark
-      val triplesRDD = spark.sparkContext.parallelize(data.getGraph.find().toList.asScala)
+      // we drop the Spark database to remove all tables from previous loaded data
+      spark.sql(s"DROP DATABASE IF EXISTS $db")
+      SparkSessionUtils.clearAllTablesAndViews(spark)
 
-      // we create a Spark database here to keep the implicit partitioning separate
-      spark.sql(s"CREATE DATABASE IF NOT EXISTS $db")
-      spark.sql(s"USE $db")
-      qef = engineFactory.create(triplesRDD)
+      // we cannot handle empty data
+      if (!data.isEmpty) {
+        // distribute on Spark
+        dataRDD = spark.sparkContext.parallelize(data.getGraph.find().toList.asScala)
+
+        // we create a Spark database here to keep the implicit partitioning separate
+        spark.sql(s"CREATE DATABASE IF NOT EXISTS $db")
+        spark.sql(s"USE $db")
+        qef = engineFactory.create(dataRDD)
+      }
+
+      data
+    } else {
+      data
     }
-
-    data
   }
 
   override def runQuery(query: Query, data: Model): SPARQLResult = {
@@ -126,4 +107,5 @@ abstract class SPARQL11TestSuiteRunnerSpark
     result
   }
 
+  super.beforeAll()
 }
